Backend/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
facenet==1.0.5
filelock==3.13.1
flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.23.4
psutil==5.9.7
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing-extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.228
urllib3==2.1.0
werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
